+++ piecrust/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `Session::set_meta` [#rusk_3341]
+- Add `Session::remove_meta` [#rusk_3341]
 
 ## [0.27.0] - 2024-12-18